requirements/dev.txt
-r test.txt

ttkbootstrap
psycopg
psycopg-binary
environs
aiofiles
inotify